Fixed subdirectory CMakeLists to properly depend on mutils
Due to the way we build each subdirectory as a separate "library"
instead of adding all the sources to one derecho library, each
subdirectory's target needs to individually declare its dependency on
the mutils::mutils target. This wasn't needed in the old system when we
had a global include_directories(mutils_INCLUDE_DIRS) declaration in the
top-level CMakeLists but now it is. Forgot to do this in my previous
commit.