Updated derechoConfig.cmake to forward public dependencies
According to the CMake documentation, any "modern" (CMake import target)
dependency that is linked publicly to a library will not automatically
be "forwarded" to downstream consumers of that library unless you write
these incantations in the Config.cmake file. This results in strange
errors like "target cascade depends on target mutils::mutils-containers
but it was not found. Are you missing a find_package() call?"