Fixed #252, improved error handling in RestartLeaderState
Added some logic to avoid the crash documented in #252 when a node that
is not in total-restart mode contacts a leader that is in total-restart
mode. The new node will send an empty view with VID -1, which will cause
the leader to treat it like a rejoining node with very out-of-date local
state (an old view and no ragged trims). Also fixed the related issue
where the leader would crash when the non-restarting node crashed
because it didn't properly handle a rejoining node disconnecting as soon
as it connected. The client_socket reads and writes in await_quorum and
receive_joiner_logs are now properly enclosed in try-catch blocks.