Replaced macros with string constants for Conf options
Using macros for constants is a holdover from C and creates problems in
C++, such as the fact that they are always in the global scope. Although
C++ historically has had problems with string constants in particular
(std::string requires dynamic allocation, so it suffers from the static
intitialization order fiasco, but using a char* where a std::string is
expected would cause extra copying), the constexpr keyword seems to
create string constants that are just as reliable and low-overhead as a
macro. A constexpr const char* can be declared and initialized in one
line and creates a compile-time constant pointer to a C-string literal,
so it's functionally equivalent to the macro but is a real named
variable that has a scope and namespace.

Note that this change requires adding the derecho:: namespace prefix in
some of the "executable" files (e.g. unit tests), so a previous
CONF_DERECHO_LOCAL_ID gets replaced with derecho::Conf::DERECHO_LOCAL_ID
instead of Conf::DERECHO_LOCAL_ID. This is a good thing, though -- it
was a little dangerous that non-Derecho code could access the CONF
macros without being in the derecho namespace.