Split config into 2 files, added option to load extra files
As a first step towards a lower-overhead config management system
(requested in #242), I changed the Conf initialization routine to expect
two files: One for the "group" config that must be the same on all
nodes, and one for a "node" config that contains only the options that
are unique per node. The "group" config is loaded first, then the "node"
config, so any options that appear in both files will take the values
specified in the "node" config file.

I also created a new top-level function, loadExtraFile, that allows a
client to repeat the config-file-loading process for an additional file
besides the two required ones. This should enable Cascade to use a
separate file for its options (i.e. "cascade.cfg") and ask Conf to load
and parse this file at startup.