Replaced volatile bool with atomic<bool> in test programs
Volatile variables are not meant to be used for synchronization between
threads and are not guaranteed to work correctly in busy-waiting loops
like we have in these tests. The correct way to ensure a variable in a
busy-waiting loop is not cached, reordered, or optimized out is to use
std::atomic, which we mostly do already in our newer test programs.