Refactored typed_subgroup_test to not use global stability callback
Apparently, as of commit eb5b281a31277faea6dd61649d057fcb094ec941 the
global stability callback is disabled for "cooked" (RPC) messages, which
means this test can't use a stability callback to detect when the last
message has been received. I'm going to try this workaround, based on
the shared-atomic-bool code I wrote for signed_store_test. I'm not sure
if it will work but I have to commit this in order to test it on
Fractus.