Improved documentation of DeserializationContext and stability callback
After talking with Weijia I think I understand how the vector of
DeserializationContext pointers can be used to link replicated objects
with state in the main thread. This is actually an important component
of making Replicated Objects work in more complicated settings, so I
added some more comments documenting what the DeserializationContexts
are used for, and replaced the confusing name "rdv" with something more
meaningful. I also updated the comments on global_stability_callback to
note that it is not called unless you use raw messages, and removed the
commented-out old code, since this design decision is not going to be
changed.