Updated signed_log_test to use DeserializationContext
This test also relied on the global stability callback to signal the
main thread, which won't work any more. Since it also used the
persistence and verified callbacks, which still work for cooked
messages, I'm creating a test-state object that receives those callbacks
as well as the "message delivered" callback, instead of putting some
logic in the main method and some within the test state object. This
also eliminates the need to create 3 counters and 3 condition variables,
since each node should only be in one subgroup, and the test object will
only contain the state for the local node's subgroup.