Moved counters into replicated state in signed_log_test
The TestState object mostly works, except that each node starts its copy
of subgroup_updates_delivered at 0, even if some updates have already
been sent in the group by the time the node joins. This can happen
because the subgroup allocator is configured with flexible subgroups, so
the Group will get an adequate configuration when the minimum number of
nodes joins, and those nodes will start sending updates even if more are
about to join. Other tests don't have this problem because they use
fixed subgroup sizes, so no updates are sent until all the nodes join,
but this test needs to include the possibility of View changes so it can
make sure the signature chains work even when the version number jumps
due to a View change. I really hope adding a non-persistent replicated
variable to each of the test objects doesn't mess anything up, but it's
the only way to allow late-joining nodes to initialize their counter to
the correct value. Otherwise those nodes wait forever for the "last
update" (even though they have received it).