Further cleanup of signed_log_test; fixed signed_bw_test
Constructor bodies should go in the cpp file too. We can use a simple
flag in the TestState object to determine whether
notify_global_persistence needs to signal the subgroup is finished,
instead of hard-coding subgroup ID 2 = UnsignedObject.

The revisions to persistent_bw_test can be easily copied to
signed_bw_test, which is almost the same except for using a signed
persistent field instead of a regular persistent field.