Updated signed_store_test to use DeserializationContext
This test was not relying on global stability callbacks, but it was
using a fragile workaround to notify the main thread when the experiment
was finished, by passing a pointer to an atomic bool into the
ObjectStore/SignatureStore factory and then hoping the subgroups were
only constructed once. Now that I know how to use the
DeserializationContext to give replicated objects pointers to the main
thread, I should do it here too, so the test can support view changes
without breaking the link between the ObjectStore objects and the main
thread.