Changed restart procedure to first attempt a normal join
If a node starts up and sees logged persistent data, it now tries to
connect to the configured contact_ip node before examining the
restart_leaders list. If the contact node responds with OK (not
TOTAL_RESTART), the joining node does not continue with the total
restart procedure and does a normal join instead. This is the same
behavior that would happen before if the restart leader responded with
OK rather than TOTAL_RESTART, but that behavior assumed that the (first)
restart leader was equal to the leader_ip (now contact_ip) which is not
necessarily true. If the contact node does not respond or crashes, the
starting node proceeds with the original total-restart procedure,
attempting to contact each node on the restart_leaders list in order.

I also refactored the handshake protocol for rejoining nodes in
total-restart mode to be more consistent with the handshake protocol in
the normal case: the joining node now sends a JoinRequest, reads a
JoinResponse, then sends its 5-tuple of port numbers in both cases. In
total-restart mode the joining node used to send its last-known-View
logs before it sent the port numbers, but now it sends them afterward.

Also changed a few comments and variable names to better distinguish
between external client sockets and "internal" (member) sockets, to make
it clearer which code paths are for external clients and which are for
joining members.