Bugfix for std::vector serialization.
The in-place deserializer for std::vector (mutils::from_bytes_no_alloc()) has a bug, where the 'const' quialifier needs to be casted off from type T when creating the context_ptr. Otherwise, it's impossible to add entries to it during deserialization. The deserializer for std::map/unordered_map has this correct.