Added debug statements and test code to try to find the bug
The "logic error in PersistentRegistry" warning gets printed in Cascade
whenever an RPC update changes only a non-signed field, and not the
signed field. This seems to indicate that the sign() method is being
called on a version that does not exist in the signed field's log, and
is trying to add a signature to this nonexistent version. I'm trying to
reproduce this bug in Derecho's signed_log_test so I can fix it here.