Further testing and logging changes to find the bug
It turns out the error only shows up when the signed persistent field is
also a DeltaSupport field, meaning it's possible for it to create a
version with no data (because the delta is empty). Regular non-Delta
signed fields always create a version with their current state even when
a non-signed field originates the makeVersion call, so mixing regular
signed and non-signed fields doesn't create a log with gaps.