Fixed sign and handle_verify_request to skip empty versions
As identified by the new signed_log_test, Delta-supporting signed
fields create a problem for the sign() method because they can end up
saving no data to the log when makeVersion is called. In particular, if
makeVersion is called when only non-signed fields were updated, the
non-signed fields will get a log entry at that version, but the
Delta-suporting signed field will just advance its current version
without creating a log entry (since there is no Delta data). Then when
PersistentRegistry::sign() is called, getMinimumVersionAfter() will
return the new version that is in the non-signed log as
cur_nonempty_version, but updateSignature won't sign any data for this
version because it doesn't exist in the signed field's log. This results
in the "logic error" message because no data was signed for
cur_nonempty_version.

The most direct solution is to simply write a new version of
getMinimumVersionAfter() that's just for sign(), called
getNextSignedVersion(), which only calls getNextVersionOf() on fields
that have signatures enabled. The getNextVersionOf() method already
correctly skips ahead to the next version with nonempty log data for
signed fields, and if no non-signed fields get to return their versions,
the overall minimum must be a version that actually has data in the log
of a signed field.

Fixing the same problem in handle_verify_request also involved fixing
a longstanding performance flaw. Instead of doing a public-key
verification on the signature posted to the SST by another replica, we
should just compare that signature to the signature in the local
replica's log on the same version - they should be bytewise identical
since both replicas have the same signing key. With this change made, we
should also ensure that we only do the equality check if both replicas
actually have a signature for the requested version; a version that is
in the persistent log but not in any signed fields doesn't need to be
verified.

Finally, this commit removes noexcept(true) from the PersistLog
constructor because the from_pem_private method (called to determine the
signature length) can throw an exception if the PEM file is not found.