Rewrote Replicated<T>::persist to separate out a "sign" method
Now that the signed_num can be different from the persisted_num
(smaller, if the current persisted_num has no signed data), it no longer
makes sense to use the single return value from Replicated::persist as
the new version that is "signed and persisted" - a single persist
request may result in a different persisted_num and signed_num.
PersistenceManager was the only part of the code calling the
Replicated::persist method, so I refactored it into separate "persist"
and "sign" methods that PersistenceManager calls in a loop, replacing
the sign-and-persist loop that used to be within Replicated::persist.
This allows PersistenceManger to learn both the latest persisted version
and the latest signed version, and update persisted_num and signed_num
separately.

While doing this, I re-arranged replicated.hpp and added some comments
to more clearly document that some of the "public" methods are in fact
only used by internal Derecho components and should not be relied on as
a stable API.

Also updated signed_log_test to correctly detect the end of the test,
now that the last persisted version might not equal the last signed
version - if the last update to MixedFieldObject is to the unsigned
field, the version resulting from that update will never be signed.