Minor fix to Derecho header includes
I noticed that a few internal files are still including other Derecho
headers with the "system" (angle-brackets) syntax instead of the "local"
(quotes) syntax. Library files should always include other library files
with the "local" syntax. Putting this in a separate commit since it's
unrelated to my other work on sign() and persist().