Changed getMinimumLatestVersion to getCurrentVersion
Following a discussion with Weijia where we confirmed that the
signatures system was the only part of Derecho that used
getMinimumLatestVersion, I decided to change this method to return the
current "in-memory" version of the Persistent fields as reported by the
currMetaHeader version, rather than the latest version of the Persistent
fields that corresponds to a log entry. I renamed the method to
getCurrentVersion to reflect this new meaning, and added a
getCurrentVersion method to the PersistLog API that would get the
meta-header version. The original getLatestVersion still exists and has
the same meaning, since the latest version in a log entry is still a
piece of information someone might want to know.