Fixed bug in handling signature requests for a version already signed
Since the sign() method, like the persist() method, can end up signing a
later version than the one requested (due to batching), it's possible
for a later persistence request to call sign() only to find that an
earlier persistence request already signed the latest version - in other
words, there's nothing for PersistentRegistry::sign() to do because the
latest version is already signed. In this case,
PersistentRegistry::sign() would return lastSignedVersion without
placing any data into the signature_buffer, because the main while loop
would do nothing. This would cause PersistenceManager to get an empty
buffer as the "latest signature" and then place that empty buffer in the
SST's signature field.

I added a special case to PersistentRegistry::sign() to make it return
the cached latest signature if the method is a no-op. I also changed
handle_persist_request to check if signed_version has actually advanced
before updating the SST, to avoid redundantly re-copying the latest
version's signature to the SST if it was already placed there by a
previous persistence request.

Also changed handle_verify_request to check the local node's version of
signed_num in the SST to determine if there's a local signature to check
against when verifying another node's signature. This is more reliable
than trying to retrieve the local signature and testing for failure,
because get_signature can "succeed" but return a signature of all zeroes
if the desired log entry has been created but not yet signed.