Rewrote persist() to persist an exact version or the latest version
The "version" argument to the various persist() methods actually didn't
do anything because FilePersistLog's persist() ignored the argument and
always persisted the latest version. After discussion with Weijia, I
rewrote FilePersistLog::persist() to actually use the argument to limit
the range of data it persists, but also to make the argument optional
and revert to the old behavior if it is not provided. This means callers
that want the batching behavior don't have to guess at the correct
"latest" version to provide as an argument (they can just provide
std::nullopt), and callers that want only a certain version persisted
will not end up persisting a later version than they wanted.

Most of the "code changes" in FilePersistLog::persist() come from me
manually inlining the NEXT_LOG_ENTRY, NEXT_LOG_ENTRY_PERS, and NEXT_DATA
macros so that I could figure out what they were doing. The substantial
change is to replace all references to m_currMetaHeader in those macros
with shadow_header, which is a modified copy of m_currMetaHeader. If the
latest_version argument is provided, shadow_header is modified to
"point" to an earlier log entry, namely the one corresponding to
latest_version. If latest_version is not provided, shadow_header equals
m_currMetaHeader and the function should behave exactly as before.
Either way, shadow_header becomes the next m_persMetaHeader after the
data is flushed, so the m_persMetaHeader accurately reflects which log
entries were persisted.