Separated verification requests into their own thread
Interleaving persistence and verification requests in a single queue
processed by a single thread makes it harder for the thread to skip
ahead and discard obsolete persistence requests when the persisted
number advances due to batching. There's also no good reason for these
requests to be handled by the same thread, since they are independent of
one another (and verification has very little to do with perisistence).