Introduce config.h to control feature selection, and ENABLE_HMEM option
This is Weijia's idea, but I'm copying it into a new commit so it can be
in a separate pull request from the Persistent API changes. The
ENABLE_HMEM option is critical to allow Derecho to run on libfabrics'
tcp provider. The config.h file puts these compile-time feature
selection macros into a header file instead of expecting them to be
passed as -D options in the compiler flags.