Updated Verbs files so they compile with the current Derecho API
Apparently we haven't tested the -DUSE_VERBS_API flag in a while,
because there are a bunch of compile errors if you use it. Some are due
to minor changes to Derecho's internals that were never propagated to
verbs.hpp/cpp, like the ip_addr_t declaration moving inside the derecho
namespace. The rest are due to the new OOB API only being implemented
on LibFabrics. I added declarations for the OOB functions to the Verbs
interface, with function bodies that simply throw exceptions, so that
Derecho can at least compile with the Verbs interface (even if none of
the OOB functions actually work).