Moved enable-timestamps macro into the new config.h
Now that there's a Cmake-generated config header for the macros that
enable or disable features, I should use it for the ENABLE_TIMESTAMPS
option. This means I can't have the macro named TIMESTAMPS (it has to
match the Cmake variable name), but that's only a minor change.

Also changed the timestamp logger files to use <> syntax for includes
instead of quotes, as we recently decided for the rest of Derecho.