Specified version limit for spdlog library
The current version of Derecho (2.4.0) was designed to compile against
older versions of the spdlog library, specifically those that came
packaged with Ubuntu 20.04 and 22.04. It turns out the spdlog library
(because of its dependency on the fmt library) made a breaking,
non-backward-compatible change in version 1.10, so this Derecho code
will stop working if spdlog is updated past version 1.9.2, which is the
version of spdlog packaged with Ubuntu 22.04. While I work on creating a
new version of Derecho that works with spdlog 1.12 (which is what is
packaged with Ubuntu 24.04), our CMakeLists should specify this version
limit for the dependency.