Set Derecho version to 2.4.1
We're ready to make the next minor release of Derecho, so this commit
will be used as the tagged commit for the release. Also fixed a minor
error in the githook script that generates git_version.cpp (the include
path should use <> not "" since it is not a relative path).