Updated README and removed an erroneous assert
The README section on configuration files was out-of-date, and still
referred to the old "leader_ip" and "leader_port" options which no
longer exist. I fixed it and added an explanation of the new "per-node
config" feature (derecho_node.cfg), which was also missing.

Also, I noticed while testing that an assert in persistence_manager
could actually fail during correct behavior, because I misunderstood the
invariants of persistence requests and persist() calls. It's possible
for persisted_version to be less than the parameter version, if the
parameter version corresponds to a null message (which does not create a
log entry). When a persistence request is posted, the argument version
is just the current version at the time the request was posted, not the
latest non-null version; thus it's not true that persist() will always
persist up to the persistence request's parameter version.