Added debug log output for derecho.cfg options
It's really hard to notice if Derecho fails to load derecho.cfg because
it will silently fall back to all-default values for the config options.
To help users debug this, I added a debug-level log statement in the
Group constructor that prints out all the actual config options that are
in memory, and changed the fallback default log level to "debug" instead
of "info" so those value will get printed out even if derecho.cfg fails
to load. I also added a console print statement to print a warning if
none of derecho.cfg, derecho_node.cfg, or command-line parameters are
present; the Conf constructor can't use Logger so console printing is
the only way to notify the user that something might be wrong.