Added files to support building Docker images for testing
Derecho shouldn't be deployed in production using Docker since it needs
direct access to the RDMA network devices, but for testing using
libfabrics' TCP adapter it might be easier to use Docker containers
connected to a virtual Docker network. This Dockerfile can be used to
build a Docker image with Derecho, either as source code in the /derecho
directory or installed as a library in /usr/local/. The compose.yaml
file can be used to launch a test network of 6 nodes that have fixed IP
addresses and sync their local copies of the Derecho code with the host
computer's derecho directory (using docker compose up --watch).