Minor fix in to_string, added a size parameter to the test
The SST to_string should unpack joiner IP addresses with inet_ntoa,
since the point of this string is to be human-readable for debugging.
The signature race condition doesn't seem to show up in signed_log_test,
just in CascadeChain, so maybe the problem is signed_log_test uses
updates that are so small they complete quickly.