Fixed initialization bug with signed_num
I discovered the source of at least one type of "Signature did not
match" warning: The signed_num field in DerechoSST was initialized to 0,
not -1, even though 0 is a valid version (the first version). This meant
handle_verify_request would try to verify another node's signature on
version 0 even if it hadn't yet put a signature in the SST signature
field. In fact, handle_verify_request needs a special case to check if a
peer node has not yet computed any signatures; it won't gracefully
degrade because 0 is a real version number and -1 is invalid as an
argument to subgroup_object->get_signature().