Partially handled race conditions when reading the signature
Since the "signature" SST field is not actually monotonic (a later
signature won't validate as an earlier signature), and signatures don't
contain their version numbers, there is actually a race condition
between a remote node updating "signature" and "signed_num" while the
local node is trying to do handle_verify_request. Since the remote node
writes in the order (1) signature (2) signed_num, while the local node
reads in the order (1) signed_num (2) signature, the local node could
read an earlier signed_num and then a later signature. This causes
signature validation to fail because the local node retrieves the wrong
signature from its log to compare (e.g. it retrieves signature 15, when
the signature it got from the SST is the signature for version 17).

This race condition can be mostly fixed by having the local node re-read
the signed_num after reading the signature and compare it with the value
it has cached. If signed_num changed while reading the signature, it
means the read interleaved with the remote node's write and we need to
try again. If signed_num has the same value twice, it's probably the
correct value (matches the signature). It is still slightly possible
to get an inconsistent read if the remote node overwrites the signature
exactly as the local node is reading it (so it won't have changed
signed_num when the local node reads it a second time), but the window
for that race condition is much smaller.