Fixed initialization of signed_num in new views; fixed default ports
Another contributor to the race condition in handle_verify_request is
the fact that the new-view initialization of the SST leaves all of the
delivery-counter fields (like persisted_num) at their default values,
assuming the first delivery predicate will update them, but
handle_verify_request can read from a remote node's signed_num before
that node has written to it. It's safer to initialize signed_num and the
corresponding signature to its correct value from the last view.
verified_num is written before it's read, like persisted_num, so it
doesn't need to be initialized.

Separately, the default values for the Derecho ports cause problems when
trying to test locally on Linux machines because they overlap with the
Linux ephemeral port range (32768 and higher). Tests keep failing
because a process accidentally connects to itself, or to another node's
system-assigned outgoing port for a different connection. It's better to
use lower values for these ports so they won't overlap with the
ephemeral range, and these values are still high enough to be unused by
any other service.